refactor: split _shared/ flat modules into rhdh_lifecycle and rhdh_prow packages
Reorganize flat modules into two proper Python packages under _shared/:

- rhdh_lifecycle/ (lifecycle data from external APIs):
  redhat.py, rhdh.py, ocp.py, pg.py
- rhdh_prow/ (openshift/release repo access):
  repo.py, yaml.py, k8s_configs.py

The two packages have zero cross-dependencies, matching the natural
boundary between lifecycle queries and CI config management.

All 15 consumer scripts updated with new import paths.
