+++ skills/_shared/rhdh_lifecycle/__init__.py
@@ -1,0 +1,1 @@
+"""RHDH lifecycle data from external APIs (Red Hat, endoflife.date)."""
