+++ skills/_shared/rhdh_lifecycle/rhdh.py
@@ -1,22 +1,20 @@
 #!/usr/bin/env python3
-"""RHDH release lifecycle data -- thin wrapper around redhat_lifecycle.
-
-Preserves the existing API surface so consumers don't need to change.
+"""RHDH release lifecycle data -- wrapper around the generic Red Hat API client.
 
 Usage:
-    from rhdh_lifecycle import fetch_rhdh_lifecycle
+    from rhdh_lifecycle.rhdh import fetch_rhdh_lifecycle
     versions = fetch_rhdh_lifecycle()
 """
 
 from __future__ import annotations
 
-from redhat_lifecycle import fetch_product_lifecycle
+from rhdh_lifecycle.redhat import fetch_api, fetch_product_lifecycle, parse_versions
 
 
 def fetch_rhdh_lifecycle(filter_version=None):
     """Fetch and parse RHDH lifecycle data."""
     versions = fetch_product_lifecycle("rhdh", filter_version)
-    # Flatten extra.ocp_versions into top-level for backward compat
+    # Flatten extra.ocp_versions into top-level for convenience
     for v in versions:
         v["ocp_versions"] = v.get("extra", {}).get("ocp_versions", [])
         v["full_support_end"] = v.get("phases", {}).get("Full support", "N/A")
@@ -25,16 +23,12 @@ def fetch_rhdh_lifecycle(filter_version=None):
 
 
 def fetch_lifecycle_api(product_name):
-    """Fetch raw API data. Delegates to redhat_lifecycle."""
-    from redhat_lifecycle import fetch_api
-
+    """Fetch raw API data. Delegates to redhat module."""
     return fetch_api(product_name)
 
 
 def parse_rhdh_versions(api_data, filter_version=None):
-    """Parse RHDH versions from raw API data. Delegates to redhat_lifecycle."""
-    from redhat_lifecycle import parse_versions
-
+    """Parse RHDH versions from raw API data."""
     versions = parse_versions(api_data, filter_version)
     for v in versions:
         v["ocp_versions"] = v.get("extra", {}).get("ocp_versions", [])
