+++ skills/_shared/rhdh_prow/__init__.py
@@ -1,0 +1,1 @@
+"""Access openshift/release repo data (CI configs, cluster pools)."""
