+++ skills/_shared/rhdh_prow/k8s_configs.py
@@ -19,11 +19,9 @@ from __future__ import annotations
 import argparse
 import re
 import sys
-from pathlib import Path
 
-sys.path.insert(0, str(Path(__file__).resolve().parent))
-from fetch_yaml import extract_branch, fetch_yaml, list_yaml_files
-from resolve_repo import resolve_repo_root
+from rhdh_prow.repo import resolve_repo_root
+from rhdh_prow.yaml import extract_branch, fetch_yaml, list_yaml_files
 
 CONFIG_DIR = "ci-operator/config/redhat-developer/rhdh"
 PREFIX = "redhat-developer-rhdh-"
