+++ skills/lifecycle-aks/scripts/check_aks_lifecycle.py
@@ -20,8 +20,8 @@ from datetime import datetime, timezone
 from pathlib import Path
 
 sys.path.insert(0, str(Path(__file__).resolve().parent.parent.parent / "_shared"))
-from fetch_yaml import print_configured_versions
-from resolve_repo import resolve_repo_root
+from rhdh_prow.repo import resolve_repo_root
+from rhdh_prow.yaml import print_configured_versions
 
 AKS_API_URL = "https://releases.aks.azure.com/parsed_data.json"
 EOL_API_URL = "https://endoflife.date/api/azure-kubernetes-service.json"
