+++ skills/lifecycle-redhat/scripts/check_lifecycle.py
@@ -21,7 +21,7 @@ import sys
 from pathlib import Path
 
 sys.path.insert(0, str(Path(__file__).resolve().parent.parent.parent / "_shared"))
-from redhat_lifecycle import (
+from rhdh_lifecycle.redhat import (
     fetch_product_lifecycle,
     list_known_products,
     resolve_product_name,
