+++ skills/lifecycle-rhdh/scripts/check_rhdh_lifecycle.py
@@ -21,7 +21,7 @@ from datetime import datetime, timezone
 from pathlib import Path
 
 sys.path.insert(0, str(Path(__file__).resolve().parent.parent.parent / "_shared"))
-from rhdh_lifecycle import fetch_rhdh_lifecycle, rhdh_supported_ocp_versions
+from rhdh_lifecycle.rhdh import fetch_rhdh_lifecycle, rhdh_supported_ocp_versions
 
 
 def main(argv=None):
