+++ skills/prow-ocp-coverage/scripts/analyze_coverage.py
@@ -21,10 +21,10 @@ from datetime import datetime, timezone
 from pathlib import Path
 
 sys.path.insert(0, str(Path(__file__).resolve().parent.parent.parent / "_shared"))
-from fetch_yaml import extract_branch, fetch_yaml, list_yaml_files
-from ocp_lifecycle import classify_ocp_versions
-from resolve_repo import resolve_repo_root
-from rhdh_lifecycle import fetch_lifecycle_api, rhdh_supported_ocp_versions
+from rhdh_lifecycle.ocp import classify_ocp_versions
+from rhdh_lifecycle.rhdh import fetch_lifecycle_api, rhdh_supported_ocp_versions
+from rhdh_prow.repo import resolve_repo_root
+from rhdh_prow.yaml import extract_branch, fetch_yaml, list_yaml_files
 
 POOL_DIR = "clusters/hosted-mgmt/hive/pools/rhdh"
 CI_CONFIG_DIR = "ci-operator/config/redhat-developer/rhdh"
@@ -107,7 +107,7 @@ def main(argv=None):
     print("--- RHDH Lifecycle ---")
     print("  Fetching from Red Hat Product Life Cycles API...")
     rhdh_response = fetch_lifecycle_api("Red Hat Developer Hub")
-    from rhdh_lifecycle import parse_rhdh_versions
+    from rhdh_lifecycle.rhdh import parse_rhdh_versions
 
     rhdh_data = parse_rhdh_versions(rhdh_response)
