+++ skills/prow-ocp-jobs/scripts/generate_test_entry.py
@@ -23,8 +23,8 @@ from io import StringIO
 from pathlib import Path
 
 sys.path.insert(0, str(Path(__file__).resolve().parent.parent.parent / "_shared"))
-from fetch_yaml import fetch_yaml, list_yaml_files
-from resolve_repo import resolve_repo_root
+from rhdh_prow.repo import resolve_repo_root
+from rhdh_prow.yaml import fetch_yaml, list_yaml_files
 from ruamel.yaml import YAML
 
 CONFIG_DIR = "ci-operator/config/redhat-developer/rhdh"
