+++ skills/prow-ocp-pools/scripts/list_cluster_pools.py
@@ -12,8 +12,8 @@ import sys
 from pathlib import Path
 
 sys.path.insert(0, str(Path(__file__).resolve().parent.parent.parent / "_shared"))
-from fetch_yaml import fetch_yaml, list_yaml_files
-from resolve_repo import resolve_repo_root
+from rhdh_prow.repo import resolve_repo_root
+from rhdh_prow.yaml import fetch_yaml, list_yaml_files
 
 POOL_DIR = "clusters/hosted-mgmt/hive/pools/rhdh"
